Added standard input redirection
