+++ src/main/java/shell/CommandLauncher.java
@@ -3,6 +3,7 @@ package shell;
 import jodd.io.StreamGobbler;
 import lombok.*;
 
+import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.IOException;
