+++ src/test/java/shell/TestCommandLauncher.java
@@ -1,12 +1,11 @@
 package shell;
 
-import java.io.File;
-import java.nio.charset.StandardCharsets;
-
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.condition.DisabledOnOs;
 import org.junit.jupiter.api.condition.EnabledOnOs;
 import org.junit.jupiter.api.condition.OS;
+
+import java.io.File;
 
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -110,49 +109,6 @@ public class TestCommandLauncher {
                 .cwd(new File(System.getProperty("java.io.tmpdir")))
                 .build();
         });
-    }
-
-    @EnabledOnOs({OS.WINDOWS})
-    @Test
-    public void redirectStdinWindows() {
-        var someText =
-                """
-                lalala
-                lolololo
-                lerelelele
-                """;
-
-        try {
-            var result = ShellCommandLauncher.builder().
-                    command("find").
-                    parameter("/c").parameter("/v").parameter("\"\"").
-                    stdin(someText.getBytes(StandardCharsets.ISO_8859_1)).build().launch();
-            assertEquals(0, result.getExitCode());
-            assertEquals("3", result.getStandardOutput().trim());
-        } catch (ShellException e) {
-            fail(e);
-        }
-    }
-
-    @DisabledOnOs({OS.WINDOWS})
-    @Test
-    public void redirectStdinUnixLike() {
-        var someText =
-                """
-                some
-                text
-                """;
-
-        try {
-            var result = ShellCommandLauncher.builder().
-                    command("wc").
-                    parameter("-l").
-                    stdin(someText.getBytes(StandardCharsets.UTF_8)).build().launch();
-            assertEquals(0, result.getExitCode());
-            assertEquals("2", result.getStandardOutput().trim());
-        } catch (ShellException e) {
-            fail(e);
-        }
     }
 
 }
